Created tutorial notebooks
