+++ scripts/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the sripts directory a Python package, allowing imports from it.
