--- sripts/model_pytorch_ViT.py
@@ -1,127 +1,0 @@
-import sys, os
-
-import torch
-from torch import nn, optim
-import timm
-import torchvision.transforms as transforms
-import torchvision.datasets as datasets
-from torch.utils.data import DataLoader
-from tqdm import tqdm
-import json
-
-
-
-data_dir = ""  # Update with your dataset path
-batch_size = 32
-num_epochs = 5
-num_classes=2
-
-device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-print(device)
-
-mean = [0.5, 0.5, 0.5]
-std = [0.5, 0.5, 0.5]
-
-data_transforms = {
-    'train': transforms.Compose([
-        transforms.Resize((224, 224)),
-        transforms.RandomHorizontalFlip(p=0.5),
-        transforms.Normalize(mean, std),
-    ]),
-    'val': transforms.Compose([
-        transforms.Resize((224, 224)),
-        transforms.RandomHorizontalFlip(p=0.5),
-        transforms.ToTensor(),
-        transforms.Normalize(mean, std),
-    ]),
-}
-
-image_datasets = {
-    x: datasets.ImageFolder(root=f"{data_dir}/{x}", transform=data_transforms[x])
-    for x in ['train', 'val']
-}
-dataloaders = {
-    x: DataLoader(image_datasets[x], batch_size=batch_size, shuffle=(x == 'train'), num_workers=4)
-    for x in ['train', 'val']
-}
-
-model_vit = timm.create_model("vit_base_patch16_224", pretrained=True, num_classes=num_classes)
-
-model_vit.to(device)
-criterion = nn.CrossEntropyLoss()
-optimizer = torch.optim.AdamW(model_vit.parameters(), lr=1e-4, weight_decay=1e-4)
-scheduler = optim.lr_scheduler.StepLR(optimizer, step_size=7, gamma=0.1)
-
-def train_model(model, dataloaders, criterion, optimizer, scheduler, num_epochs, save_path="training_progress.json"):
-    history = {"train_loss": [], "val_loss": [], "train_acc": [], "val_acc": []}
-
-    for epoch in range(num_epochs):
-        print(f"Epoch {epoch+1}/{num_epochs}")
-        print("-" * 10)
-
-        for phase in ['train', 'val']:
-            if phase == 'train':
-                model.train()
-            else:
-                model.eval()
-
-            running_loss = 0.0
-            running_corrects = 0
-
-            for inputs, labels in tqdm(dataloaders[phase]):
-                inputs, labels = inputs.to(device), labels.to(device)
-                optimizer.zero_grad()
-
-                with torch.set_grad_enabled(phase == 'train'):
-                    outputs = model(inputs)
-                    _, preds = torch.max(outputs, 1)
-                    loss = criterion(outputs, labels)
-
-                    if phase == 'train':
-                        loss.backward()
-                        optimizer.step()
-
-                running_loss += loss.item() * inputs.size(0)
-                running_corrects += torch.sum(preds == labels.data)
-
-            if phase == 'train' and scheduler is not None:
-                scheduler.step()
-
-            epoch_loss = running_loss / len(dataloaders[phase].dataset)
-            epoch_acc = running_corrects.double() / len(dataloaders[phase].dataset)
-
-            print(f"{phase} Loss: {epoch_loss:.4f} Acc: {epoch_acc:.4f}")
-
-            history[f"{phase}_loss"].append(epoch_loss)
-            history[f"{phase}_acc"].append(epoch_acc.item())
-
-        with open(save_path, "w") as f:
-            json.dump(history, f)
-
-    return model
-
-def evaluate_model(model, dataloader):
-    model.eval()
-    correct = 0
-    total = 0
-    y_true, y_pred = [], []
-
-    with torch.no_grad():
-        for inputs, labels in dataloader:
-            inputs, labels = inputs.to(device), labels.to(device)
-            outputs = model(inputs)
-            _, preds = torch.max(outputs, 1)
-            correct += (preds == labels).sum().item()
-            total += labels.size(0)
-
-            y_true.extend(labels.cpu().numpy())
-            y_pred.extend(preds.cpu().numpy())
-
-    print(f"Validation Accuracy: {correct / total:.4f}")
-
-
-trained_model = train_model(model_vit, dataloaders, criterion, optimizer, scheduler, num_epochs)
-
-torch.save(trained_model.state_dict(), "model_ViT.pth")
-
-evaluate_model(trained_model, dataloaders['val'])
